backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
numpy==1.26.4
opencv-python==4.10.0.84
scipy==1.13.1
python-socketio==5.11.4

# RDK runtime dependencies are provided by board BSP:
# hobot_dnn
# hobot_vio
